avcenc: Prepend startcodes to output buffers
This makes the encoder output data that is similar to actual HW
encoders.

This also makes the encoder's output work in other decoders
when passing through directly. If feeding the encoded data
straight through to a decoder (such as in the EncodeDecodeTest
CTS test), the output from OMX.google.h264.encoder cannot be
decoded by e.g. OMX.Nvidia.h264.decode on a Nexus 7 or
OMX.SEC.avc.dec on a Galaxy S3. (The first frame can be somewhat
decoded by the sw decoder OMX.google.h264.decoder, but not properly.)

Change-Id: I6e16433a4a66936e3080723c8233681f53bd55e5